+++ QmBase-Shared/QmBase-Cache/src/main/java/com/qm/base/shared/cache/core/annotation/QmCacheable.java
@@ -24,7 +24,7 @@ public @interface QmCacheable {
     /**
      * 缓存过期时间（单位秒），300 表示使用默认 ttl
      */
-    int ttl() default 300;
+    long ttl() default 300;
 
     /**
      * 是否缓存 null 值，默认为 false
